# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dilshod's API Gateway",
        "title": "API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/exchange/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new exchange record with the specified details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "Create a new Exchange",
                "parameters": [
                    {
                        "enum": [
                            "buy",
                            "sell"
                        ],
                        "type": "string",
                        "description": "Exchange Status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Exchange details",
                        "name": "exchange",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.CreateExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange created successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while creating exchange",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/exchange/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific exchange record using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "Delete Exchange by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exchange ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ExchangeResponse"
                        }
                    },
                    "404": {
                        "description": "Exchange not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while deleting exchange",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/exchange/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific exchange record using its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "Get Exchange by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exchange ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetExchangeResponse"
                        }
                    },
                    "404": {
                        "description": "Exchange not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while retrieving exchange",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/exchange/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a list of all exchange records with optional filtering by product ID or status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "List all Exchanges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "buy",
                            "sell"
                        ],
                        "type": "string",
                        "description": "Exchange Status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of exchanges retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.GetAllExchangeResponse"
                        }
                    },
                    "500": {
                        "description": "Error while listing exchanges",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/exchange/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing exchange record with new details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange"
                ],
                "summary": "Update an Exchange",
                "parameters": [
                    {
                        "description": "Exchange details",
                        "name": "exchange",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genprotos.UpdateExchangeRequest"
                        }
                    },
                    {
                        "enum": [
                            "buy",
                            "sell"
                        ],
                        "type": "string",
                        "description": "Exchange Status",
                        "name": "status",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange updated successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.ExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Exchange not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while updating exchange",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/minio/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a file to MinIO",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MinIO"
                ],
                "summary": "Upload File",
                "parameters": [
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Filename",
                        "name": "filename",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File uploaded successfully",
                        "schema": {
                            "$ref": "#/definitions/genprotos.UploadFileResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Error while uploading file",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "genprotos.CreateExchangeRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "description": "Enum: buy, sell",
                    "type": "string"
                }
            }
        },
...
        "genprotos.ExchangeResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "genprotos.GetAllExchangeResponse": {
            "type": "object",
            "properties": {
                "all_exchanges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genprotos.GetExchangeResponse"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.GetExchangeResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.UpdateExchangeRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "contract_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "genprotos.UploadFileResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}